+++ packages/api/src/modules/redis/redis.service.ts
@@ -4,7 +4,7 @@ import { forEach } from 'p-iteration';
 import { WINSTON_MODULE_PROVIDER } from 'nest-winston';
 import { Logger } from 'winston';
 
-import { REDIS_CONFIG } from 'src/config';
+import { REDIS_CONFIG, DEBUG_REDIS } from 'src/config';
 import { CacheKeys } from './cache.dto';
 
 @Injectable()
@@ -19,30 +19,27 @@ export class RedisService {
       port: REDIS_CONFIG.port,
       password: REDIS_CONFIG.password,
     });
-
-    // clear cache when api starts
-    this.clearCache();
   }
 
-  private clearCache() {
-    this.logger.info('clear cache');
+  public clearCache() {
+    if (DEBUG_REDIS) this.logger.info('clear cache');
     return forEach(Object.entries(CacheKeys), ([, value]) =>
       this.deleteKeysPattern(value)
     );
   }
 
   public get(key: string) {
-    this.logger.info('get key', { key });
+    if (DEBUG_REDIS) this.logger.info('get key', { key });
     return this.client.get(key);
   }
 
   public set(key: string, data: string, ttl: number) {
-    this.logger.info('set key', { key });
+    if (DEBUG_REDIS) this.logger.info('set key', { key });
     return this.client.set(key, data, 'PX', ttl);
   }
 
   public async deleteKeysPattern(key: CacheKeys) {
-    this.logger.info('delete key', { key });
+    if (DEBUG_REDIS) this.logger.info('delete key', { key });
     const keys = await this.client.keys(`${key}*`);
     await Promise.all(keys.map((_) => this.client.del(_)));
   }