+++ packages/web/components/calandar/calendar.styles.tsx
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const CalendarStyles = styled.div`
+  .wrapper {
+    padding-top: 60px;
+    max-width: 1200px;
+    margin: 0 auto;
+  }
+
+  .cell-title {
+    font-weight: 600;
+  }
+`;