+++ packages/web/utils/graphql.tsx
@@ -125,6 +125,12 @@ export type JackettInput = {
   tag: Scalars['String'];
 };
 
+export type LibraryCalendar = {
+   __typename?: 'LibraryCalendar';
+  movies: Array<EnrichedMovie>;
+  tvEpisodes: Array<EnrichedTvEpisode>;
+};
+
 export type Movie = {
    __typename?: 'Movie';
   id: Scalars['Float'];
@@ -139,6 +145,7 @@ export type Mutation = {
    __typename?: 'Mutation';
   saveQualityParams: GraphQlCommonResponse;
   saveTags: GraphQlCommonResponse;
+  clearRedisCache: GraphQlCommonResponse;
   updateParams: GraphQlCommonResponse;
   startScanLibraryJob: GraphQlCommonResponse;
   startFindNewEpisodesJob: GraphQlCommonResponse;
@@ -260,6 +267,7 @@ export type Query = {
   getMissingTVEpisodes: Array<EnrichedTvEpisode>;
   getMissingMovies: Array<EnrichedMovie>;
   getTVSeasonDetails: Array<EnrichedTvEpisode>;
+  getCalendar: LibraryCalendar;
   omdbSearch: OmdbInfo;
 };
 
@@ -436,6 +444,17 @@ export type UpdateParamsInput = {
   value: Scalars['String'];
 };
 
+export type ClearCacheMutationVariables = {};
+
+
+export type ClearCacheMutation = (
+  { __typename?: 'Mutation' }
+  & { result: (
+    { __typename?: 'GraphQLCommonResponse' }
+    & Pick<GraphQlCommonResponse, 'success' | 'message'>
+  ) }
+);
+
 export type StartScanLibraryMutationVariables = {};
 
 
@@ -601,6 +620,27 @@ export type UpdateParamsMutation = (
   & { result: (
     { __typename?: 'GraphQLCommonResponse' }
     & Pick<GraphQlCommonResponse, 'success' | 'message'>
+  ) }
+);
+
+export type GetCalendarQueryVariables = {};
+
+
+export type GetCalendarQuery = (
+  { __typename?: 'Query' }
+  & { calendar: (
+    { __typename?: 'LibraryCalendar' }
+    & { movies: Array<(
+      { __typename?: 'EnrichedMovie' }
+      & Pick<EnrichedMovie, 'id' | 'title' | 'state' | 'releaseDate'>
+    )>, tvEpisodes: Array<(
+      { __typename?: 'EnrichedTVEpisode' }
+      & Pick<EnrichedTvEpisode, 'id' | 'episodeNumber' | 'seasonNumber' | 'state' | 'releaseDate'>
+      & { tvShow: (
+        { __typename?: 'TVShow' }
+        & Pick<TvShow, 'id' | 'title'>
+      ) }
+    )> }
   ) }
 );
 
@@ -867,6 +907,38 @@ export type SearchQuery = (
 );
 
 
+export const ClearCacheDocument = gql`
+    mutation clearCache {
+  result: clearRedisCache {
+    success
+    message
+  }
+}
+    `;
+export type ClearCacheMutationFn = ApolloReactCommon.MutationFunction<ClearCacheMutation, ClearCacheMutationVariables>;
+
+/**
+ * __useClearCacheMutation__
+ *
+ * To run a mutation, you first call `useClearCacheMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useClearCacheMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [clearCacheMutation, { data, loading, error }] = useClearCacheMutation({
+ *   variables: {
+ *   },
+ * });
+ */
+export function useClearCacheMutation(baseOptions?: ApolloReactHooks.MutationHookOptions<ClearCacheMutation, ClearCacheMutationVariables>) {
+        return ApolloReactHooks.useMutation<ClearCacheMutation, ClearCacheMutationVariables>(ClearCacheDocument, baseOptions);
+      }
+export type ClearCacheMutationHookResult = ReturnType<typeof useClearCacheMutation>;
+export type ClearCacheMutationResult = ApolloReactCommon.MutationResult<ClearCacheMutation>;
+export type ClearCacheMutationOptions = ApolloReactCommon.BaseMutationOptions<ClearCacheMutation, ClearCacheMutationVariables>;
 export const StartScanLibraryDocument = gql`
     mutation startScanLibrary {
   result: startScanLibraryJob {
@@ -1296,6 +1368,54 @@ export function useUpdateParamsMutation(baseOptions?: ApolloReactHooks.MutationH
 export type UpdateParamsMutationHookResult = ReturnType<typeof useUpdateParamsMutation>;
 export type UpdateParamsMutationResult = ApolloReactCommon.MutationResult<UpdateParamsMutation>;
 export type UpdateParamsMutationOptions = ApolloReactCommon.BaseMutationOptions<UpdateParamsMutation, UpdateParamsMutationVariables>;
+export const GetCalendarDocument = gql`
+    query getCalendar {
+  calendar: getCalendar {
+    movies {
+      id
+      title
+      state
+      releaseDate
+    }
+    tvEpisodes {
+      id
+      tvShow {
+        id
+        title
+      }
+      episodeNumber
+      seasonNumber
+      state
+      releaseDate
+    }
+  }
+}
+    `;
+
+/**
+ * __useGetCalendarQuery__
+ *
+ * To run a query within a React component, call `useGetCalendarQuery` and pass it any options that fit your needs.
+ * When your component renders, `useGetCalendarQuery` returns an object from Apollo Client that contains loading, error, and data properties 
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useGetCalendarQuery({
+ *   variables: {
+ *   },
+ * });
+ */
+export function useGetCalendarQuery(baseOptions?: ApolloReactHooks.QueryHookOptions<GetCalendarQuery, GetCalendarQueryVariables>) {
+        return ApolloReactHooks.useQuery<GetCalendarQuery, GetCalendarQueryVariables>(GetCalendarDocument, baseOptions);
+      }
+export function useGetCalendarLazyQuery(baseOptions?: ApolloReactHooks.LazyQueryHookOptions<GetCalendarQuery, GetCalendarQueryVariables>) {
+          return ApolloReactHooks.useLazyQuery<GetCalendarQuery, GetCalendarQueryVariables>(GetCalendarDocument, baseOptions);
+        }
+export type GetCalendarQueryHookResult = ReturnType<typeof useGetCalendarQuery>;
+export type GetCalendarLazyQueryHookResult = ReturnType<typeof useGetCalendarLazyQuery>;
+export type GetCalendarQueryResult = ApolloReactCommon.QueryResult<GetCalendarQuery, GetCalendarQueryVariables>;
 export const GetDiscoverDocument = gql`
     query getDiscover($entertainment: Entertainment, $originLanguage: String, $primaryReleaseYear: String, $score: Float, $genres: [Float!], $page: Float) {
   TMDBResults: discover(entertainment: $entertainment, originLanguage: $originLanguage, primaryReleaseYear: $primaryReleaseYear, score: $score, genres: $genres, page: $page) {